requirements.txt
# LangChain ecosystem (upgraded from 0.1.x to 0.3.x)
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langchain-text-splitters>=0.3.0,<0.4.0

# Graph orchestration (upgraded from 0.0.48 to 0.6+)
langgraph>=0.6.0,<0.7.0

# Web framework
fastapi>=0.110.0,<1.0.0
uvicorn>=0.27.0,<1.0.0
sse-starlette

# LLM & Search
google-search-results>=2.4.2,<3.0.0

# Vector search
faiss-cpu

# Auth & Config
python-jose[cryptography]
python-dotenv>=1.0.1,<2.0.0

# Type system (using Python 3.11+ native typing)
typing_extensions>=4.11.0,<5.0.0
pydantic>=2.0.0,<3.0.0
